subsystem 3/app/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
requests